sources/requirements.txt
typer    == 0.15.1
RPi.GPIO == 0.7.1
